Fix cyclic dependency
series.ts needs to know about `SeriesOptionModule` (file: module.ts), but
the `Module` type (file: module.ts) needs to know about the `Series` type
(file: series.ts). Therefore split the module.ts into two files:

-   coreModules.ts (axis, legend, series).
-   optionModules.ts (axis-options, series-options).

Currently these are all in util/, but I will probably move these into a
common directory somewhere else soon.
