+++ packages/ag-charts-community/src/chart/agChartV2.ts
@@ -25,21 +25,16 @@ import {
     type SeriesOptionsTypes,
 } from './mapping/types';
 import { windowValue } from '../util/window';
-import {
-    REGISTERED_MODULES,
-    type AxisOptionModule,
-    type LegendModule,
-    type Module,
-    type ModuleInstance,
-    type RootModule,
-    type SeriesOptionModule,
-} from '../util/module';
 import { Logger } from '../util/logger';
 import { getJsonApplyOptions } from './chartOptions';
 
 import { setupModules } from './factory/setupModules';
 import { getLegendKeys } from './factory/legendTypes';
 import { registerInbuiltModules } from './factory/registerInbuiltModules';
+import { type Module, REGISTERED_MODULES } from '../util/module';
+import type { ModuleInstance } from '../util/baseModule';
+import type { LegendModule, RootModule } from '../util/coreModules';
+import type { AxisOptionModule, SeriesOptionModule } from '../util/optionModules';
 
 type ProcessedOptions = Partial<AgChartOptions> & { type?: SeriesOptionsTypes['type'] };
