+++ packages/ag-charts-community/src/chart/factory/axisTypes.ts
@@ -1,5 +1,5 @@
-import type { AxisConstructor } from '../../util/module';
 import type { ModuleContext } from '../../util/moduleContext';
+import type { AxisConstructor } from '../../util/coreModules';
 import { LogAxis } from '../axis/logAxis';
 import { NumberAxis } from '../axis/numberAxis';
 import { CategoryAxis } from '../axis/categoryAxis';
