+++ packages/ag-charts-community/src/chart/factory/seriesTypes.ts
@@ -3,7 +3,7 @@ import { PieSeries } from '../series/polar/pieSeries';
 import { TreemapSeries } from '../series/hierarchy/treemapSeries';
 import type { ChartType } from './chartTypes';
 import { registerChartSeriesType } from './chartTypes';
-import type { SeriesConstructor, SeriesPaletteFactory } from '../../util/module';
+import type { SeriesConstructor, SeriesPaletteFactory } from '../../util/coreModules';
 import type { ModuleContext } from '../../util/moduleContext';
 import type { AgChartOptions } from '../../options/agChartOptions';
 import type { SeriesOptionsTypes } from '../mapping/types';
