+++ packages/ag-charts-community/src/module-support.ts
@@ -2,6 +2,9 @@ export * from './util/array';
 export * from './util/validation';
 export * from './util/default';
 export { extractDecoratedProperties, isDecoratedObject, listDecoratedProperties } from './util/decorator';
+export * from './util/baseModule';
+export * from './util/coreModules';
+export * from './util/optionModules';
 export * from './util/module';
 export * from './util/moduleContext';
 export * from './util/navigator-module';
